refactor(wasm): split api/styling.rs into color + prepass modules
styling.rs (581 LOC) bundled two unrelated concerns under one "styling" name:
the colour find/extract/resolve chain and the EntityScanner pre-pass + job
collection + site rotation. Convert it into an api/styling/ directory with
color.rs (colour chain + its 5 tests) and prepass.rs (PrePassData,
combined_pre_pass, type-geometry job collection, building rotation), behind a
mod.rs that re-exports the six pub(crate) items siblings reference.

Pure code motion (0 lines changed; verbatim). Visibility preserved exactly;
the two private colour helpers stay private; the 3 pre-existing dead-code
colour fns still warn (warning set byte-identical to baseline). api/mod.rs
untouched. cargo test -p ifc-lite-wasm green (12); wasm32 build clean.
